+++ frontend/api/_lib/types/category.type.ts
@@ -1,8 +1,9 @@
-export type CategoryType = "Income" | "Expense" | "Financial"
+export type CategoryType = "Income" | "Expense" | "System"
 
 export type Category = {
   id: string
   name: string
   type: CategoryType
   parentId: string | null
+  note: string
 }